Move ingredient getting middleware to separate folder
