+++ src/routes/generate.js
@@ -1,44 +1,11 @@
 const express = require('express')
-const mongoose = require('mongoose')
-const Joi = require('joi')
-const ingredientSchema = require('../schemas/ingredient')
-
-mongoose.connect('mongodb://127.0.0.1:27017/tiktaco')
-  .then(() => console.log('Connected to MongoDB...'))
-  .catch((err) => console.log('Could not connect to MongoDB', err))
-
-const BaseLayer = mongoose.model('Base Layer', ingredientSchema)
-const Condiment = mongoose.model('Condiment', ingredientSchema)
-const Mixin = mongoose.model('Mixin', ingredientSchema)
-const Seasoning = mongoose.model('Seasoning', ingredientSchema)
-const Shell = mongoose.model('Shell', ingredientSchema)
-
-const getIngredients = async (req, res, next) => {
-  if (req.method !== 'GET') {
-    next()
-  }
-  
-  const baseLayer = await getRandomIngredient(BaseLayer)
-  const condiment = await getRandomIngredient(Condiment)
-  const mixin = await getRandomIngredient(Mixin)
-  const seasoning = await getRandomIngredient(Seasoning)
-  const shell = await getRandomIngredient(Shell)
-
-  console.log({ baseLayer, condiment, mixin, seasoning, shell })
-  next()
-}
-
-const getRandomIngredient = async (model) => {
-  const ingredients = await model.find({})
-  return ingredients[Math.floor(Math.random() * ingredients.length)]
-}
-
+const getIngredients = require('../middleware/aggregator')
 const router = express.Router()
 
-router.use(getIngredients)
+router.use('/', getIngredients)
 
 router.get('/', (req, res) => {
-  res.send('Hello world!')
+  res.send(req.taco)
 })
 
 module.exports = router
